% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestNeighbors.R
\name{findNearestNeighbors}
\alias{findNearestNeighbors}
\title{Find the approximate set of neighbors of each voxels.}
\usage{
findNearestNeighbors(maskImg, radius = 1, threads = parallel::detectCores())
}
\arguments{
\item{maskImg}{Input Masked image which the masked voxels are labeled with zeros}

\item{radius}{Integer: Indices that `dist(targetInd - queryInd) <= radius`}

\item{threads}{Integer: threads to use in closest point search.}
}
\value{
The result from \code{\link{vcgKDtree}}
}
\description{
Find the approximate set of neighborhood of each voxels,
given the distance radius.The number of elements in each neighborhood is (2*radius)^3.
We recommend to find the nearest neighbors prior to \code{limma}
}
