% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_SE.R
\name{nifti_df_to_SummarizedExperiment}
\alias{nifti_df_to_SummarizedExperiment}
\title{Convert NIfTI data.frame to a SummarizedExperiment}
\usage{
nifti_df_to_SummarizedExperiment(
  x,
  path_column = "image",
  assay_name = path_column,
  template_images = NULL,
  ...
)
}
\arguments{
\item{x}{An optional `data.frame` describing the samples/participants.}

\item{path_column}{column in `x` that is the paths to the file
names of the data.  Can also be a list column of `x` where this column
are the images}

\item{assay_name}{Name of the `assay` in the
\code{\link{SummarizedExperiment}}}

\item{template_images}{List of images that are annotations of the rows
of the data (voxels), such as template mask or segmentations}

\item{...}{additional arguments to pass to \code{\link{SummarizedExperiment}}}
}
\value{
A \code{SummarizedExperiment} output
}
\description{
Convert NIfTI data.frame to a SummarizedExperiment
}
\examples{
tarfile = system.file("extdata", "can.tar.gz", package = "limmi")
tarfile
exdir = tempfile()
dir.create(exdir)
files = untar(tarfile = tarfile, list = TRUE, exdir = exdir)
files = files[!grepl("^[.]", basename(files))]
unz = untar(tarfile = tarfile, files = files, exdir = exdir)
files = files[ grepl("hdr", basename(files))]
files
files = file.path(exdir, files)
df = data.frame(image = files,
age = stats::rpois(length(files), 50))
se = nifti_df_to_SummarizedExperiment(df, "image")
}
